Moved some fields from Advanced to Standard insight
